requirements.txt
streamlit==1.27.2
pandas
pdfplumber
tabula-py

